Refactor and improve the docs
